constructor <-> ingredients
