+++ src/components/burger-constructor/burger-constructor.jsx
@@ -1,14 +1,22 @@
 import React from 'react';
-import style from './burger-constructor.module.css';
-import Tabs from './tabs';
+import {ConstructorElement} from '@ya.praktikum/react-developer-burger-ui-components';
 
-export default class BurgerConstructor extends React.Component{
+export default class BurgerIngredients extends React.Component{
   render() {
     return (
-    <section className={'container'} style={{backgroundColor: 'red'}}>
-      <h2 className={style.header}>Соберите бургер</h2>
-      <Tabs />
-      {'Constructor!'}
+    <section className={'container'}>
+      {
+        this.props.ingredients.map(el=>
+          < ConstructorElement
+            // type={'top'}
+            isLocked={false}
+            // handleClose?: () => void;
+            text ={el.name}
+            thumbnail={el.image}
+            price={el.price}
+          />
+        )}
+      {'Hello!'}
     </section>
   )}
 }
