--- src/components/burger-constructor/burger-constructor.module.css
@@ -1,3 +1,0 @@
-.header{
-  margin: 40px 0 20px 0;
-}
