+++ src/components/burger-ingredients/burger-ingredients.jsx
@@ -1,22 +1,14 @@
 import React from 'react';
-import {ConstructorElement} from '@ya.praktikum/react-developer-burger-ui-components';
+import style from './burger-ingredients.module.css';
+import Tabs from './tabs';
 
-export default class BurgerIngredients extends React.Component{
+export default class BurgerConstructor extends React.Component{
   render() {
     return (
-    <section className={'container'}>
-      {
-        this.props.ingredients.map(el=>
-          < ConstructorElement
-            // type={'top'}
-            isLocked={false}
-            // handleClose?: () => void;
-            text ={el.name}
-            thumbnail={el.image}
-            price={el.price}
-          />
-        )}
-      {'Hello!'}
+    <section className={'container'} style={{backgroundColor: 'red'}}>
+      <h2 className={style.header}>Соберите бургер</h2>
+      <Tabs />
+      {'Constructor!'}
     </section>
   )}
 }
