+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,0 +1,3 @@
+.header{
+  margin: 40px 0 20px 0;
+}
